+++ CHANGES.rst
@@ -9,6 +9,7 @@ in progress
 - Add two more examples about using `explore dashboards` with `jq`
 - CI: Prepare test suite for testing two different dashboard schema versions, v27 and v33
 - Improve determinism by returning stable sort order of dashboard results
+- Improve compatibility with Grafana 8.3 by handling dashboard schema version 33 properly
 
 2021-12-11 0.12.0
 =================